Patch onOpen._add function in tinyMCE.
Because the update to 1.3.10 the javascript throws errors when attaching
an event handler to ```onOpen``` because the ```_add``` function is not
in the scope of Event but the ```add``` function. So the best
workaround it to bind the ```_add``` function to ```add```.i
See https://github.com/4teamwork/bern.web/pull/206 for update
information.