Safari and ie10 displays border under each navigation point.
Because of diffrent calculating of line-height predefined from the body
the navigation elements does not have equal height. So use display block
to get rid of this influence. Because of the same reason use px
instead of em for setting padding on navigation elements. Set fixed
height to make it displayed equal for each browser.