load style.top inline CSS nodes early.
By moving style.top nodes before other link / style / script nodes,
the inline CSS of those nodes is rendered early.
This helps us to avoid flickering of certain components, such as
background images, by inlining the CSS and caching the resources
in the browser.

The style-tag needs to have the "top" class, so that we do not mess with
the order of other inlined resources by accident.